requirements.txt
netifaces
flask
readline
pycrypto
